refactor: 일부 코드 controller에서 view로 이동
