+++ src/main/java/lotto/view/View.java
@@ -2,7 +2,13 @@ package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import lotto.common.constants.GuideMessage;
+import lotto.common.constants.LottoDefaultRule;
+import lotto.common.constants.LottoRankRule;
+import lotto.common.constants.Rank;
 import lotto.common.constants.Symbol;
 import lotto.common.utils.Utils;
 import lotto.common.validate.Validate;
@@ -52,5 +58,63 @@ public class View {
     public void prizeStatsMessage() {
         System.out.println(GuideMessage.OUTPUT_PRIZE_STATS.getMessage());
         System.out.println(Symbol.DIVIDE_LINE.getSymbol());
+    }
+
+    Map<String, Integer> matchRecords = new HashMap<String, Integer>() {{
+        put(Rank.FIRST_RANK.getRank(), 0);
+        put(Rank.SECOND_RANK.getRank(), 0);
+        put(Rank.THIRD_RANK.getRank(), 0);
+        put(Rank.FOURTH_RANK.getRank(), 0);
+        put(Rank.FIFTH_RANK.getRank(), 0);
+    }};
+
+    public void compareTicketsAndLotto(ArrayList<ArrayList<Integer>> ticketsNumbers, List<Integer> lottoNumbers, int bonusNumber) {
+        ticketsNumbers.forEach(ticketNumbers -> {
+            int numberMatchCount = ticketNumberAndLottoNumberMatchCount(ticketNumbers, lottoNumbers);
+            int ticketBonusNumber = useBonusNumber(ticketNumbers);
+            rankResult(numberMatchCount, ticketBonusNumber, bonusNumber);
+        });
+        System.out.println(matchRecords);
+    }
+
+    private int ticketNumberAndLottoNumberMatchCount(List<Integer> ticketNumbers, List<Integer> lottoNumbers) {
+        ticketNumbers.removeAll(lottoNumbers);
+        return LottoDefaultRule.PICK_HIT_NUMBER_TOTAL.getRule() - ticketNumbers.size();
+    }
+
+    private int useBonusNumber(List<Integer> ticketNumbers) {
+        if (ticketNumbers.size() == 1) {
+            return ticketNumbers.get(0);
+        }
+        return 0;
+    }
+
+    private boolean rankResult(int numberMatchCount, int ticketNumberOnlyOne, int bonusNumber) {
+        if (numberMatchCount == LottoRankRule.FIRST_RANK_MATCH_COUNT.getRank()) {
+            matchRecords.put(Rank.FIRST_RANK.getRank(), matchRecords.get(Rank.FIRST_RANK.getRank()) + 1);
+            return true;
+        }
+        if (numberMatchCount == LottoRankRule.SECOND_RANK_MATCH_COUNT.getRank()) {
+            isMatchTicketAndBonus(ticketNumberOnlyOne, bonusNumber);
+            return true;
+        }
+        if (numberMatchCount == LottoRankRule.FOURTH_RANK_MATCH_COUNT.getRank()) {
+            matchRecords.put(Rank.FOURTH_RANK.getRank(), matchRecords.get(Rank.FOURTH_RANK.getRank()) + 1);
+            return true;
+        }
+        if (numberMatchCount == LottoRankRule.FIFTH_RANK_MATCH_COUNT.getRank()) {
+            matchRecords.put(Rank.FIFTH_RANK.getRank(), matchRecords.get(Rank.FIFTH_RANK.getRank()) + 1);
+            return true;
+        }
+        return true;
+    }
+
+    private boolean isMatchTicketAndBonus(int ticketNumber, int bonusNumber) { // 번호가 5개 일치할 경우 진입
+        if (ticketNumber == bonusNumber) {
+            matchRecords.put(Rank.SECOND_RANK.getRank(), matchRecords.get(Rank.SECOND_RANK.getRank()) + 1);
+            return true;
+        }
+        matchRecords.put(Rank.THIRD_RANK.getRank(), matchRecords.get(Rank.THIRD_RANK.getRank()) + 1);
+        return true;
     }
 }
